refactor(sync): move sync infrastructure into infra packages
Organize the PR's sync code under 'infra' to match core/data's existing infra
pattern (NetworkMonitor, TimeZoneMonitor, …), instead of the module root / util:

- core/data:      Synchronizer + Syncable + NetworkChange + snapshotSync/changeListSync
                  (kpt.core.data -> kpt.core.data.infra); SyncManager (kpt.core.data.util
                  -> kpt.core.data.infra); the 4 sync tests -> kpt.core.data.infra.
- core/datastore: ChangeListVersions + SyncStatePersister/SettingsSyncStatePersister
                  (kpt.core.datastore -> kpt.core.datastore.infra).

git mv (history preserved) + imports updated across all consumers (core/data,
core/datastore DI, sync worker, feature test fakes). No behavior change.
Verified: main + test compile across all affected modules.
