+++ core/data/src/commonMain/kotlin/kpt/core/data/economic/MacroIndicatorsRepository.kt
@@ -13,7 +13,7 @@ import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.Flow
 import kpt.core.base.store.screen.FetchPolicy
 import kpt.core.base.store.screen.ScreenDataStream
-import kpt.core.data.Syncable
+import kpt.core.data.infra.Syncable
 import kpt.core.model.economic.MacroIndicator
 import kpt.core.store.economic.impl.MacroIndicatorKey
