+++ core/datastore/src/commonMain/kotlin/kpt/core/datastore/di/DatastoreModule.kt
@@ -12,10 +12,10 @@ package kpt.core.datastore.di
 import com.russhwolf.settings.Settings
 import kpt.core.base.common.di.CommonModule
 import kpt.core.base.datastore.di.DatastoreBaseModule
-import kpt.core.datastore.SettingsSyncStatePersister
-import kpt.core.datastore.SyncStatePersister
 import kpt.core.datastore.UserPreferencesRepository
 import kpt.core.datastore.UserPreferencesRepositoryImpl
+import kpt.core.datastore.infra.SettingsSyncStatePersister
+import kpt.core.datastore.infra.SyncStatePersister
 import org.koin.core.qualifier.named
 import org.koin.dsl.bind
 import org.koin.dsl.module
