+++ feature/home/src/commonTest/kotlin/kpt/feature/home/ui/HomeDashboardViewModelTest.kt
@@ -27,11 +27,11 @@ import kpt.core.base.store.screen.FetchPolicy
 import kpt.core.base.store.screen.ScreenDataStream
 import kpt.core.base.store.screen.ScreenState
 import kpt.core.base.store.screen.screenDataStreamForTesting
-import kpt.core.data.Synchronizer
 import kpt.core.data.banking.BillReminderRepository
 import kpt.core.data.banking.LoanRepository
 import kpt.core.data.currency.CurrencyRepository
 import kpt.core.data.economic.EconomicRatesRepository
+import kpt.core.data.infra.Synchronizer
 import kpt.core.model.banking.BillReminder
 import kpt.core.model.banking.Loan
 import kpt.core.model.currency.ExchangeRates
