+++ feature/home/src/commonTest/kotlin/kpt/feature/home/ui/HomeViewModelTest.kt
@@ -28,11 +28,11 @@ import kpt.core.base.store.screen.FetchPolicy
 import kpt.core.base.store.screen.ScreenDataStream
 import kpt.core.base.store.screen.ScreenState
 import kpt.core.base.store.screen.screenDataStreamForTesting
-import kpt.core.data.Synchronizer
 import kpt.core.data.banking.BillReminderRepository
 import kpt.core.data.banking.LoanRepository
 import kpt.core.data.currency.CurrencyRepository
 import kpt.core.data.economic.EconomicRatesRepository
+import kpt.core.data.infra.Synchronizer
 import kpt.core.model.banking.BillCategory
 import kpt.core.model.banking.BillReminder
 import kpt.core.model.banking.Loan
